added condition to see if display exist

+++ post_install_script.sh
@@ -4,23 +4,28 @@ echo "updating"
 sudo apt update; sudo apt upgrade -y
 
 
-sudo apt install i3
 
 sudo apt install tmux
 
 sudo apt install git
 
-sudo apt install gimp
+
+# Check if display exist
+if [ -n "${DISPLAY+x}" ]; then 
+	sudo apt install i3
+
+	sudo apt install gimp
 
 
-# Installing latest chrome
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
 
-sudo apt install ./google-chrome-stable_current_amd64.deb
+	# Installing latest chrome
+	wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
 
-rm google-chrome-stable_current_amd64.deb
+	sudo apt install ./google-chrome-stable_current_amd64.deb
 
+	rm google-chrome-stable_current_amd64.deb
 
+fi
 
 ## Instaing latest nvim and allow for global access of the command "nvim"
 # Also adds nvim dependencies
